pyproject.toml
[tool.poetry]
name = "booking"
version = "0.1.0"
description = ""
authors = ["Pavel Ivanov <pavel.ivanov.060988@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
cloudinary = "^1.39.0"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
